Change Navigation structure and usage, fix bug
