+++ package.json
@@ -4,7 +4,10 @@
   "description": "Shared Component Library for Bandwidth React Apps",
   "main": "dist/index.js",
   "scripts": {
-    "watch": "babel --watch --copy-files --out-dir dist src",
+    "watch:clean": "rimraf dist",
+    "watch:run": "babel --watch --copy-files --out-dir dist src",
+    "watch": "run-s watch:*",
+    "build:clean": "rimraf dist",
     "build:indices": "node tools/generateIndices.js",
     "build:dist": "babel --copy-files --out-dir dist src",
     "build": "run-s build:*",
@@ -43,6 +46,7 @@
     "babel-preset-stage-0": "^6.24.1",
     "file": "^0.2.2",
     "lodash": "^4.17.4",
-    "npm-run-all": "^4.0.2"
+    "npm-run-all": "^4.0.2",
+    "rimraf": "^2.6.1"
   }
 }
