+++ src/index.js
@@ -43,8 +43,8 @@ export { default as Modal } from './components/Modal';
 export { default as Money } from './components/Money';
 export { default as Navigation } from './components/Navigation';
 export { default as LogoHeader } from './components/Navigation/LogoHeader';
-export { default as NavigationHeader } from './components/Navigation/NavigationHeader';
 export { default as NavigationItem } from './components/Navigation/NavigationItem';
+export { default as NavigationItems } from './components/Navigation/NavigationItems';
 export { default as NewBadge } from './components/NewBadge';
 export { default as Pagination } from './components/Pagination';
 export { default as Radio } from './components/Radio';
